Mini Project/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
delta-spark==4.0.0
idna==3.10
importlib_metadata==8.7.0
numpy==2.3.1
pandas==2.3.0
pillow==11.3.0
py4j==0.10.9.9
pyarrow==20.0.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-mnist==0.7
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
